backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
aiofiles==23.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
email-validator==2.3.0
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
slowapi==0.1.9

